network_cmd_port - buildfix

+++ configuration.h
@@ -306,7 +306,7 @@ typedef struct settings
    bool savestate_auto_load;
 
    bool network_cmd_enable;
-   uint16_t network_cmd_port;
+   unsigned network_cmd_port;
    bool stdin_cmd_enable;
 
    char core_assets_directory[PATH_MAX_LENGTH];
